-Corregidos errores en función modificar rutinas. -Añadida busqueda a fragmento de ejercicios.

+++ app/src/main/java/com/unex/proyectoasee_nogymmembership/Adds/AddExerciseToRoutine.java
@@ -112,7 +112,7 @@ public class AddExerciseToRoutine extends AppCompatActivity implements RoutineLi
         protected Exercise doInBackground(Void... voids) {
 
             AppDataBase appDB = AppDataBase.getDataBase(AddExerciseToRoutine.this);
-            Exercise exercise=appDB.exerciseDAO().getExercise(id_ex);
+            Exercise exercise=appDB.exerciseDAO().getExercise((int)id_ex);
 
 
             if(exercise==null){
